css/css-animations/parsing/animation-name-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-name with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-name">
<meta name="assert" content="animation-name supports the full grammar '[ none | <keyframes-name> ]#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("animation-name", 'NONE', 'none');

test_valid_value("animation-name", 'foo');
test_valid_value("animation-name", 'Both');
test_valid_value("animation-name", 'ease-in');
test_valid_value("animation-name", 'infinite');
test_valid_value("animation-name", 'paused');
test_valid_value("animation-name", 'first, second, third');

test_valid_value("animation-name", '"string"', ['"string"', "string"]);
test_valid_value("animation-name", '"multi word string"', ['"multi word string"', "multi\\ word\\ string"]);
test_valid_value("animation-name", '"initial"', ['"initial"', "initial"]);
test_valid_value("animation-name", '"---\\22---"', ['\"---\\\"---\"', '---\\\"---']);

test_valid_value("animation-name", '"none"', ['"none"', "none"]);
test_valid_value("animation-name", '"inherit"', ['"inherit"', "inherit"]);
test_valid_value("animation-name", '"unset"', ['"unset"', "unset"]);
test_valid_value("animation-name", '"default"', ['"default"', "default"]);
test_valid_value("animation-name", '"revert"', ['"revert"', "revert"]);
test_valid_value("animation-name", '"revert-layer"', ['"revert-layer"', "revert-layer"]);
</script>
</body>
</html>
